donkies/donkies/requirements/development.txt
celery==4.0.0
Django==1.10.3
django-cors-headers==1.3.0
django-debug-toolbar==1.6
django-filter==0.15.3
django-oauth-toolkit==0.10.0
djangorestframework==3.5.3
factory-boy==2.7.0
fake-factory==0.7.2
Pillow==3.4.2
psycopg2==2.6.2
pytest==3.0.3
pytest-django==3.0.0
python-social-auth==0.2.21
redis==2.10.5
requests==2.11.1
uWSGI==2.0.14
